+++ src/audio_stream_opus_chunked.h
@@ -62,7 +62,7 @@ protected:
 
 public:
     virtual int32_t _mix_resampled(AudioFrame *dst_buffer, int32_t frame_count) override;
-    virtual double _get_stream_sampling_rate() const override;
+    virtual float _get_stream_sampling_rate() const override;
 
     virtual void _start(double p_from_pos = 0.0) override;
     virtual void _stop() override;
